+++ data/src/main/java/in/koreatech/koin/data/mapper/RecruitmentMapper.kt
@@ -1,10 +1,6 @@
-﻿@file:Suppress("TooManyFunctions")
-
-package `in`.koreatech.koin.data.mapper
+﻿package `in`.koreatech.koin.data.mapper
 
 import `in`.koreatech.koin.data.request.recruitment.RecruitmentUpdateRequest
-import `in`.koreatech.koin.data.request.recruitment.TeamRecruitmentActivityRequest
-import `in`.koreatech.koin.data.request.recruitment.TeamRecruitmentRoleRequest
 import `in`.koreatech.koin.data.response.recruitment.MyAppliedRecruitmentResponse
 import `in`.koreatech.koin.data.response.recruitment.MyRecruitmentPostResponse
 import `in`.koreatech.koin.data.response.recruitment.RecruitmentApplicationResponse
@@ -14,10 +10,6 @@ import `in`.koreatech.koin.data.response.recruitment.RecruitmentNotificationList
 import `in`.koreatech.koin.data.response.recruitment.RecruitmentNotificationResponse
 import `in`.koreatech.koin.data.response.recruitment.RecruitmentResponse
 import `in`.koreatech.koin.data.response.recruitment.RecruitmentRoleResponse
-import `in`.koreatech.koin.data.response.recruitment.TeamRecruitmentActivityResponse
-import `in`.koreatech.koin.data.response.recruitment.TeamRecruitmentApplicationResponse
-import `in`.koreatech.koin.data.response.recruitment.TeamRecruitmentApplicationRoleResponse
-import `in`.koreatech.koin.data.response.recruitment.TeamRecruitmentProfileResponse
 import `in`.koreatech.koin.domain.model.recruitment.MyAppliedRecruitment
 import `in`.koreatech.koin.domain.model.recruitment.MyRecruitmentPost
 import `in`.koreatech.koin.domain.model.recruitment.Recruitment
@@ -28,12 +20,6 @@ import `in`.koreatech.koin.domain.model.recruitment.RecruitmentNotifications
 import `in`.koreatech.koin.domain.model.recruitment.RecruitmentRole
 import `in`.koreatech.koin.domain.model.recruitment.RecruitmentUpdate
 import `in`.koreatech.koin.domain.model.recruitment.Recruitments
-import `in`.koreatech.koin.domain.model.recruitment.TeamRecruitmentActivity
-import `in`.koreatech.koin.domain.model.recruitment.TeamRecruitmentActivityInput
-import `in`.koreatech.koin.domain.model.recruitment.TeamRecruitmentApplication
-import `in`.koreatech.koin.domain.model.recruitment.TeamRecruitmentApplicationRole
-import `in`.koreatech.koin.domain.model.recruitment.TeamRecruitmentProfile
-import `in`.koreatech.koin.domain.model.recruitment.TeamRecruitmentRoleInput
 
 fun MyRecruitmentPostResponse.toMyRecruitmentPost() = MyRecruitmentPost(
     id = id,
@@ -171,50 +157,4 @@ fun RecruitmentUpdate.toRecruitmentUpdateRequest() = RecruitmentUpdateRequest(
     description = description,
     relatedUrl = relatedUrl,
     qualification = qualification
-)
-
-fun TeamRecruitmentProfileResponse.toTeamRecruitmentProfile() = TeamRecruitmentProfile(
-    profileNickname = profileNickname,
-    department = department,
-    major = major,
-    studentNumber = studentNumber,
-    preferredRole = preferredRole,
-    skills = skills,
-    activities = activities.map { it.toTeamRecruitmentActivity() },
-    selfIntroduction = selfIntroduction
-)
-
-fun TeamRecruitmentActivityResponse.toTeamRecruitmentActivity() = TeamRecruitmentActivity(
-    id = id,
-    title = title,
-    startedAt = startedAt,
-    endedAt = endedAt,
-    isOngoing = isOngoing,
-    description = description
-)
-
-fun TeamRecruitmentActivityInput.toTeamRecruitmentActivityRequest() = TeamRecruitmentActivityRequest(
-    title = title,
-    startedAt = startedAt,
-    endedAt = endedAt,
-    isOngoing = isOngoing,
-    description = description
-)
-
-fun TeamRecruitmentRoleInput.toTeamRecruitmentRoleRequest() = TeamRecruitmentRoleRequest(
-    name = name,
-    maxParticipants = maxParticipants
-)
-
-fun TeamRecruitmentApplicationResponse.toTeamRecruitmentApplication() = TeamRecruitmentApplication(
-    applicationId = applicationId,
-    recruitmentId = recruitmentId,
-    status = status,
-    role = role.toTeamRecruitmentApplicationRole(),
-    createdAt = createdAt
-)
-
-fun TeamRecruitmentApplicationRoleResponse.toTeamRecruitmentApplicationRole() = TeamRecruitmentApplicationRole(
-    id = id,
-    name = name
 )
